FEATURE Now Makefile .o rules depends on .h
Makefile was edited to make a dependency between the header file and
the compilation of object files. This way a change in the header file
forces the re-compilation of object files.

Also the 42 header was corrected on ft_lst*.c files because it was
still showing the old file name (i.e. _bonus.c).