ADDED ft_atoi_b() function FIXED Makefile
A version of atoi for base capabilities was added. ft_atoi_b()
can convert positive numbers in a string for bases from 2 to 16.

Makefile was updated and fixed: The rule to print the "building"
title was forcing the execution of the rule that had it as a
prerequisite. This happened because every rule that doesn't
generate a file target forces other rules that have it as a
prerequisite. This was fixed by making the "building" title rule
to touch a file that serves as a flag. This file is deleted
along the cleaning of the objects.

Documentation and TODO were updated.