ADDED ft_free_strarr_from()
In the collection ft_strarr.h the function ft_free_strarr_from()
was added to free allocated memory from a certain string in the
array to the last one. When the index is 0 this function behaves
just as ft_free_strarr() thus also liberating the allocated memory
for the argv style NULL terminated c-string array.