FIXED Makefile, README.md re-creation
A trick was added to be able to re-create the README.md from the
comments in a 42 norm .h header file. This capability was added
as a pre-requisite in the Makefile when building the documentation.

After limiting the output when building libft which needed to use
a flag file to signal the completion of the printing of the building
title. The deletion of the auxiliary file "print_title" is now done
at "fclean" rule. This was neccesary because when libft is added as a
submodule in another project, when fcleaning that project the libft
clean rule is "remotely" executed first and after that the project
Makefile also remotely executes the libft fclean rule.
So when the deletion of "print_title" was originally performed at
libft clean rule, the later remote execution of libft fclean rule
throws an error because the file was already removed by the prior
execution of libft clean rule as it is prerequisite of libft
clean rule.

The .gitignore was also edited to never stage the auxiliary file
"print_title" in order to avoid accidental inconsistencies in
projects that use libft as a submodule. If the case arise that
libft repository were updated after libft.a had been locally built,
commiting the auxiliary file will leave the submodule in an
inconsistent state: libft.a has not been yet built for the project
that needs it but the "print_title" auxiliary file exists (which
should only happen after a libft.a building).